<⚡️ perf>[client]:{reduce call component and fix multi-language}
